Add files via upload
This is example of using highcharts. However I had to create my own data. I created my data in a python file and via a http request I used that data in my immediately invoked function. The data came in as an object, so I made an array. This was done to easily apply the data to the highcharts. Highcharts.js take data in different formats.  Anyways read through the code. You will see ways of sorting and mapping arrays, dynamically creating tables from data that is turned into percentage. Have fun